firewall: use proxmox_log
The firewall did not yet use proxmox-log as its logging provider.
Adapt the existing code to setup logging via proxmox-log over the
current custom solution and import proxmox-log in all files as an
alias to the log cate.

This commit removes the ability to debug proxmox-firewall via running
it in the foreground and dumping the log output, since it now always
gets written to the journal. This will be fixed in a future commit,
that introduces dedicated subcommands for debugging.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Reviewed-by: Gabriel Goller <g.goller@proxmox.com>
Tested-by: Gabriel Goller <g.goller@proxmox.com>
Link: https://lore.proxmox.com/20250527135800.190084-2-s.hanreich@proxmox.com
 [TL: resolve context merge conflict and use trixie's proxmox-log 1]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>