fix #6107: add support for legacy ipset / alias names
Change the existing firewall rule generation logic to accomodate the
changes in the config parser, so it can deal with the legacy ipset /
alias names.

The lookup logic is the same as in pve-firewall: In guest context,
check if it is contained in the guest config, otherwise check the
cluster config. In cluster context, only check the cluster
configuration.

Because the rule generation logic now has to look up ipsets instead of
blindly inserting rules with the ipset name, store the generated SDN
ipsets in the firewall config as well, to avoid re-generating them for
every lookup. This should also allow for better error reporting, in
case ipsets or aliases are referenced that do not exist in the
configuration.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250925122403.230867-4-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>