fix #6336: fix ipfilter matching logic
Matching on ipsets in the firewall generally works by matching on two
sets (one for match, one for nomatch):

  ip saddr @ipfilter ip saddr != @ipfilter-nomatch <verdict>

Ipfilters were created with the comparison operators simply inverted,
which leads to ipfilters with empty nomatch sets never working, since
the second expression always evaluates to false on empty sets:

  ip saddr != @ipfilter ip saddr = @ipfilter-nomatch drop

In order to properly invert the logic, two statements need to be
generated (analogous to negation in boolean logic):

  ip saddr != @ipfilter drop
  ip saddr = @ipfilter-nomatch drop

To avoid cluttering the existing set handling function and to clearly
separate the rule generation logic, introduce a completely new
function that handles ipfilters separately. This also allows the set
handling function to validate the name of ipsets in a later commit,
since it only operates on ipsets that have to exist in the firewall
config, whereas ipfilters do not necessarily exist in the firewall
configuration.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251001162818.320717-5-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>